% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setcwd.R
\name{setcwd}
\alias{setcwd}
\title{Set current working directory}
\usage{
setcwd(directory = NULL)
}
\arguments{
\item{directory}{Optional. A directory path to set as the working directory.}
}
\description{
Set current working directory to the directory of the active script in RStudio, or to a specified directory if provided.
}
\examples{
\dontrun{
setcwd()
}
}
