% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewmarkdown.R
\name{viewmarkdown}
\alias{viewmarkdown}
\title{Send Markdown to RStudio Viewer Pane}
\usage{
viewmarkdown(markdown)
}
\arguments{
\item{markdown}{Markdown variable to be rendered on the Viewer Pane}
}
\description{
Send Markdown to RStudio Viewer Pane
Frequently, the Markdown will be the output of an LLM request
}
