% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaymedia.R
\name{displaymedia}
\alias{displaymedia}
\title{Display media file}
\usage{
displaymedia(filename = "")
}
\arguments{
\item{filename}{The filename, either Local or URL}
}
\description{
Display media file on Rstudio, either on "Plots" or "Viewer" panels
Supported formats: jpg, jpeg, png, webm, mp4, pdf, youtube
Local files or Cloud are supported
}
\examples{
\dontrun{
displaymedia() # Displays all Local media files
displaymedia("https://rstudio.github.io/cheatsheets/rstudio-ide.pdf")
displaymedia("https://www.w3schools.com/tags/movie.mp4")
displaymedia("https://www.youtube.com/watch?v=uXIIk7suD6c")
displaymedia("sample.mp4")
displaymedia("sample.pdf")
}
}
